requirements.txt
mysql-connector-python
requests
pandas
SQLAlchemy
psycopg2